# [Perf][MTP] Improve speculative decoding(MTP) efficiency